Fix crash: db is null on database failure (M68 merge)
(cherry picked from commit f88ccb0851d0259fec2a4e37d59d0ba3caa1b99d)

Bug: 860228
Reviewed-on: https://chromium-review.googlesource.com/1129466
Commit-Queue: Dan H <harringtond@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#575105}
Reviewed-on: https://chromium-review.googlesource.com/1140736
Cr-Commit-Position: refs/branch-heads/3440@{#695}
Cr-Branched-From: 010ddcfda246975d194964ccf20038ebbdec6084-refs/heads/master@{#561733}

+++ components/offline_pages/core/model/cleanup_thumbnails_task.cc
@@ -18,7 +18,9 @@ typedef base::OnceCallback<void(CleanupThumbnailsTask::Result)> ResultCallback;
 
 CleanupThumbnailsTask::Result CleanupThumbnailsSync(base::Time now,
                                                     sql::Connection* db) {
-  const char kSql[] =
+  if (!db)
+    return CleanupThumbnailsTask::Result();
+  static const char kSql[] =
       "DELETE FROM page_thumbnails "
       "WHERE offline_id IN ("
       "  SELECT pt.offline_id from page_thumbnails pt"
